fixed reprojection error calculation
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@11796 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ vision/checkerboard_detector/checkerboard_calibration.cpp
@@ -229,7 +229,7 @@ private:
                     cvReleaseMat(&new_image_points);
                     off += _vNumPointsPerImage[i];
                 }
-                err = sqrt(err)/_vAllPoints.size();
+                err = sqrt(err/(double)_vAllPoints.size());
 
                 ROS_INFO("calibration done, reprojection error = %f", (float)err);
                 ROS_INFO("Intrinsic Matrix:");
